feat(#1398): CLEAN: Split init.go (946 nbnc) by responsibility
